2023 day 24, part 2, copied from reddit
This is a math problem to solve a system of equations with 6
variables, so I don't know how to do that. Looks like the sympy python
module helps here a lot because it solves equation systems but I just
copied the equations from reddit, I didn't do any math here.